Move fetch* functions to lib
