+++ src/index.ts
@@ -1,5 +1,3 @@
-export { default as fetchStringNode } from './fetchStringNode';
-export { default as fetchStringPuppeteer } from './fetchStringPuppeteer';
 export { default as findFavicons } from './findFavicons';
 export { default as getFavicons } from './getFavicons';
 export * from './constants';
